backend/requirements.txt
requests
Flask
Flask-Cors
psycopg2-binary
cpe
OTXv2
